[feat] MAC4-17 로그아웃 기능 구현 (#10)
