+++ src/main/java/com/mac4/yeopabackend/user/controller/UserController.java
@@ -1,0 +1,23 @@
+package com.mac4.yeopabackend.user.controller;
+
+import com.mac4.yeopabackend.common.response.ApiResponse;
+import com.mac4.yeopabackend.user.service.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/user")
+@RequiredArgsConstructor
+public class UserController {
+
+    private final UserService userService;
+
+    @PostMapping("/logout")
+    public ApiResponse<Void> logout(@RequestHeader("Authorization") String authorization){
+        userService.logout(authorization);
+        return ApiResponse.success();
+    }
+}
